Fix chrome build (pem file check)

+++ build.sh
@@ -15,9 +15,13 @@ mv chas-correct.xpi ../bin/chas-correct-firefox-mobile.xpi
 cd ..
 
 #Для хрома
-./crxmake.sh chrome ../correct.pem
-mv chrome.crx bin/chas-correct-chrome-desktop.crx
-cp bin/chas-correct-chrome-desktop.crx bin/chas-correct.zip
+if [ -f chrome.pem ]; then
+  ./crxmake.sh chrome chrome.pem
+  mv chrome.crx bin/chas-correct-chrome-desktop.crx
+  cp bin/chas-correct-chrome-desktop.crx bin/chas-correct.zip
+else
+  echo '"chrome.pem" file not found. Skipping Chrome ext build.'
+fi
 
 #По просьбам трудящихся - юзерскрипт
 cat jquery-2.1.0.min.js jstorage.min.js dictionary.js prepareDictionary.js content.js > bin/chas-correct-userscript.js
